fix(core): :zap: component movement and cleanup
component movement and cleanup

Ref #109
