+++ app/components.registry.js
@@ -1,22 +1,22 @@
 import Components from "./components/component-docs/Components";
-import SampleLayouts from "./components/component-docs/layouts/SampleLayouts";
+import BlankLayoutPage from "./components/layout-docs/page/BlankLayoutPage";
+import CenteredBlankLayoutPage from "./components/layout-docs/page/CenteredBlankLayoutPage";
+import ComplexLayoutPage from "./components/layout-docs/page/ComplexLayoutPage";
+import FixedFooterLayoutPage from "./components/layout-docs/page/FixedFooterLayoutPage";
+import FixedHeaderFooterLayoutPage from "./components/layout-docs/page/FixedHeaderFooterLayoutPage";
+import FixedHeaderLayoutPage from "./components/layout-docs/page/FixedHeaderLayoutPage";
+import FooterLayoutPage from "./components/layout-docs/page/FooterLayoutPage";
+import HCenteredBlankLayoutPage from "./components/layout-docs/page/HCenteredBlankLayoutPage";
+import HeaderFooterLayoutPage from "./components/layout-docs/page/HeaderFooterLayoutPage";
+import HeaderLayoutPage from "./components/layout-docs/page/HeaderLayoutPage";
+import LeftDrawerLayoutPage from "./components/layout-docs/page/LeftDrawerLayoutPage";
+import LeftRightDrawerPage from "./components/layout-docs/page/LeftRightDrawerPage";
+import RightDrawerLayoutPage from "./components/layout-docs/page/RightDrawerLayoutPage";
+import ThreeColumnLayoutPage from "./components/layout-docs/page/ThreeColumnLayoutPage";
+import TwoColumnLayoutPage from "./components/layout-docs/page/TwoColumnLayoutPage";
+import VCenteredBlankLayoutPage from "./components/layout-docs/page/VCenteredBlankLayoutPage";
+import SampleLayouts from "./components/layout-docs/SampleLayouts";
 import StyleUtilities from "./components/styles-docs/StyleUtilities";
-import BlankLayoutPage from "./layout/page/BlankLayoutPage";
-import CenteredBlankLayoutPage from "./layout/page/CenteredBlankLayoutPage";
-import ComplexLayoutPage from "./layout/page/ComplexLayoutPage";
-import FixedFooterLayoutPage from "./layout/page/FixedFooterLayoutPage";
-import FixedHeaderFooterLayoutPage from "./layout/page/FixedHeaderFooterLayoutPage";
-import FixedHeaderLayoutPage from "./layout/page/FixedHeaderLayoutPage";
-import FooterLayoutPage from "./layout/page/FooterLayoutPage";
-import HCenteredBlankLayoutPage from "./layout/page/HCenteredBlankLayoutPage";
-import HeaderFooterLayoutPage from "./layout/page/HeaderFooterLayoutPage";
-import HeaderLayoutPage from "./layout/page/HeaderLayoutPage";
-import LeftDrawerLayoutPage from "./layout/page/LeftDrawerLayoutPage";
-import LeftRightDrawerPage from "./layout/page/LeftRightDrawerPage";
-import RightDrawerLayoutPage from "./layout/page/RightDrawerLayoutPage";
-import ThreeColumnLayoutPage from "./layout/page/ThreeColumnLayoutPage";
-import TwoColumnLayoutPage from "./layout/page/TwoColumnLayoutPage";
-import VCenteredBlankLayoutPage from "./layout/page/VCenteredBlankLayoutPage";
 
 export const ComponentsRegistry = {
   BlankLayoutPage            : { comp: BlankLayoutPage },
