+++ app/components/layout-docs/page/FixedFooterLayoutPage.js
