+++ app/components/layout-docs/page/FooterLayoutPage.js
