+++ app/components/layout-docs/page/TwoColumnLayoutPage.js
