+++ app/components/layout-docs/page/VCenteredBlankLayoutPage.js
